Feat: executeDelete
1. executePut과 마찬가지로 file에 대해서만 적용함. rmDir을 적용 할 수 도 있을 것 같으나,
현재에서는 URI을 보고 파일인지 아닌지 판단하고, 모든 메소드는 파일에 적용하는 방식을 사용하고 있기 때문에, 배제됨.

+++ srcs/Server_execute*.cpp
@@ -94,10 +94,50 @@ void		Server::executePut(Connection& connection, const Request& request, std::st
 
 void		Server::executeDelete(Connection& connection, const Request& request, std::string target_uri)
 {
-	connection.set_m_response(new Response(&connection, 200));
+	bool file_exist = 0;
+	int open_fd = open(target_uri.c_str(), O_RDWR);
+	int temp = errno;
+	if (open_fd > 0)
+	{
+		file_exist = true;
+		if (errno == 0) // NOTE 파일를 open으로 열었음
+		{
+			int status_code = 200;
+			connection.set_m_response(new Response(&connection, status_code, request.getBody()));
+			std::string errorpage_body = Response::makeErrorPage(status_code, connection.get_m_request()->GetMethod());
+			connection.get_m_response()->set_m_body(errorpage_body);
+			unlink(target_uri.c_str());
+			temp = errno;
+		}
+		else if (errno == EISDIR) // NOTE 폴더를 open으로 열었음
+		{
+			int status_code = 204;
+			connection.set_m_response(new Response(&connection, status_code));
+			rmdir(target_uri.c_str());
+			temp = errno;
+		}
+		close(open_fd);
+		temp = errno;
+	}
+	else // ENOENT 2
+	{
+		int status_code = 204;
+		connection.set_m_response(new Response(&connection, status_code));
+		int open_fd = open(target_uri.c_str(), O_WRONLY | O_CREAT, 0755);
+		write(open_fd, connection.get_m_response()->get_m_body().c_str(), connection.get_m_response()->get_m_body().length());
+		close(open_fd);
+	}
 	Response *response = connection.get_m_response();
-	int ret = unlink(target_uri.c_str());
-	cout << "ret: " << ret << endl;
+
+	response->set_m_headers("Date", ft::getCurrentTime().c_str());
+	response->set_m_headers("Server", "webserv");
+	if (errno == 0 && file_exist == true)
+	{
+		response->set_m_headers("Content-Length", ft::itoa(response->get_m_body().length()));
+		response->set_m_headers("Content-Type", "text/html");
+		response->set_m_headers("Content-Language", "en-US");
+	}
+	errno = 0; // NOTE 초기화!
 }
 
 void		Server::executeOptions(Connection& connection, const Request& request)
